+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/asset.ftl
@@ -1,1 +1,1 @@
-${dataFactory.toInsertSQL(assetDataFactory.newDefaultAssetVocabularyModel())}
+${resourcePermissionDataFactory.toInsertSQL(assetDataFactory.newDefaultAssetVocabularyModel())}
