+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/counters.ftl
@@ -4,6 +4,6 @@
 	<#if '${counterModel.name}' == 'com.liferay.counter.kernel.model.Counter'>
 		update Counter set currentId = ${counterModel.currentId} where name = '${counterModel.name}';
 	<#else>
-		${dataFactory.toInsertSQL(counterModel)}
+		${resourcePermissionDataFactory.toInsertSQL(counterModel)}
 	</#if>
 </#list>
