+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/roles.ftl
@@ -1,3 +1,3 @@
 <#list userDataFactory.newRoleModels(classNameDataFactory.getClassNameId("Role")) as roleModel>
-	${dataFactory.toInsertSQL(roleModel)}
+	${resourcePermissionDataFactory.toInsertSQL(roleModel)}
 </#list>
